Added /disease route
This is my solution to a problem I though of.
When we will have routing for different diseases we'll want our url to be /disease/{diseaseName}
this will cause problem as we render MainPage ONLY when we match '/' on url (it has the exact prop set true)
so now, when the url is /disease/{diseaseName} it will still render the MainPage.

We'll need to implement disease own links and more testing or possible changes.