Can switch between login and signup
Removed for now login and signup routes as we should use different components to render as the style is a bit different, and if we have a route we should make it atlesat look better so we'll make our own in another issue.

Added open signUp and openSignIn functions to be passed to the signIn/Up forms.

Changed the Links to Typograpghy and added uniqe styling.