CR fix + DiseasePage Fix
Changed namings
Added the same functonlity to the disease page.

Future solution we need to discuss is either an AbstractMainPage that MainPage and DiseasePage (SitePage in the future) will made of, or something else that wraps them both, or we stay with the same architercure as we'll differ more in the future.

anyway, I don't think it's this branch related.